Scripts/requirements.txt
# Python Dependencies for Model Download and Conversion

# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
ultralytics>=8.0.0

# Data Processing and Image Handling
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0

# Kaggle API
kaggle>=1.5.0

# Additional utilities
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For advanced model optimization
onnx>=1.12.0
onnxruntime>=1.12.0
